pyproject.toml
[project]
name = "voice-synth-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.10"
dependencies = [
    "flask>=3.1.0",
    "numpy>=1.22.0",
    "openai>=1.56.2",
    "openai-whisper",
    "setuptools>=75.6.0",
    "speechrecognition>=3.11.0",
    "tts>=0.22.0",
    "python-dotenv>=1.0.1",
    "flask-cors>=5.0.0",
    "pyaudio>=0.2.14",
]

[tool.uv.sources]
openai-whisper = { git = "https://github.com/openai/whisper.git" }
